fix flatten with empty lists/tuples/dicts, other minor tweaks
